Move api calling logic to dogs/index.jsx, conditionally show CardDogTable
